1.3 driver doc
Suggesting to 
1) add a section for how to pass in routing context from the driver
2) include a section for Kerberos auth token

The Kerberos auth token requires example code from all drivers. The example will be tagged with `kerberos-auth`
The newly added docs requires proper links in the doc to give more information to explain the new features added in driver, as the driver just expose interfaces to the new features added on the server side.